Added initComputedStyle to avoid deriving layout values from unstyled elements.
In some cases, such as bootstrap modals and popovers, the directive may be linked prior to being inserted into the document. In this state, there are no styles applied to the element. When this occurred, even emitting elastic:adjust did not recalculate the layout values. With this change, the directive is initialized properly without the elastic:adjust, but if cases exist where it is not, emitting elastic:adjust will calculate the layout if styles were not previously available.